requirements.txt
pytest==5.2.1
python-telegram-bot==12.2.0
python-dotenv==0.10.3
